apistrandsagentproject1/requirements.txt
flask==3.0.0
flask-cors==4.0.0
strands-agents>=1.0.0
python-dotenv==1.0.0
boto3==1.34.22
httpx==0.27.2
llm-guard
